Kill wpa_supplicant before load driver
When system_server restarts due to crash,
the start sequence for wifi is as follows:
1.) unload wifi driver
2.) load wifi driver
3.) kill wpa_supplicant that was left
4.) start new wpa_supplicant

The problem is that when we load the wifi driver,
the old wpa_supplicant will be notified immediatedly
due to netlink, and tries to bring up the interface
This is unnecessary.

We should kill wpa_supplicant before loading wifi driver
to avoid unnecessary bring up of interface by old wpa_supplicant
and possibly race condition in kernel driver.


Change-Id: Ie39c9b847d811e09c55386c1c4cf5f554af05868